Map Automatisierung, Kamera Verfolgung, Neues Player Movement

+++ game/objects/tiles/Air.java
@@ -4,13 +4,13 @@ import java.awt.Graphics2D;
 
 public class Air extends Tile{
 
-	public Air(int x, int y) {
+	public Air(double x, double y) {
 		super(x, y);
 	}
 
 	@Override
 	public void render(Graphics2D g, double tileSize) {
-		//ignore
+		//ignorieren (nicht sichtbar)
 	}
 
 }
